+++ dotnet/Data/CybersourceRepository.cs
@@ -68,6 +68,18 @@
                 if (response.IsSuccessStatusCode)
                 {
                     merchantSettings = JsonConvert.DeserializeObject<MerchantSettings>(responseContent);
+                    /// TESTING
+                    //if (merchantSettings.MerchantDefinedValues == null)
+                    //{
+                    //    merchantSettings.MerchantDefinedValues = new Dictionary<int, string>();
+                    //    merchantSettings.MerchantDefinedValues.Add(1, "{{MerchantId}}");        // MID
+                    //    merchantSettings.MerchantDefinedValues.Add(2, "{{CompanyName}}");       // Order Type
+                    //    merchantSettings.MerchantDefinedValues.Add(3, "{{CompanyTaxId}}");      // CALL CENTER
+                    //    merchantSettings.MerchantDefinedValues.Add(4, "{{MiniCart.Buyer.LastName}},{{MiniCart.Buyer.FirstName}}");      // Customer Name
+                    //    merchantSettings.MerchantDefinedValues.Add(5, "{{TotalCartValue}}");    // Total Cart Amount
+                    //    //await this.SetMerchantSettings(merchantSettings);
+                    //}
+                    /// TESTING
                 }
             }
             catch(Exception ex)
@@ -384,7 +396,6 @@
             var client = _clientFactory.CreateClient();
             var response = await client.SendAsync(request);
             string responseContent = await response.Content.ReadAsStringAsync();
-            Console.WriteLine($"GetOrderConfiguration [{response.StatusCode}] ");
             if (response.StatusCode == HttpStatusCode.NotFound)
             {
                 return null;