+++ dotnet/Data/ICybersourceRepository.cs
@@ -1,6 +1,4 @@
 ﻿using Cybersource.Models;
-using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Cybersource.Data